requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
